7120 - Ensure trial session is updated when adding case to hearing
