+++ shared/src/business/test/createTestApplicationContext.js
@@ -337,7 +337,7 @@ const createTestApplicationContext = ({ user } = {}) => {
   });
 
   const mockGetPersistenceGateway = appContextProxy({
-    addHearingToCase: jest.fn(),
+    addCaseToHearing: jest.fn(),
     addWorkItemToSectionInbox,
     bulkDeleteRecords: jest.fn().mockImplementation(bulkDeleteRecords),
     bulkIndexRecords: jest.fn().mockImplementation(bulkIndexRecords),
