+++ shared/src/business/useCases/trialSessions/setForHearingInteractor.test.js
@@ -87,10 +87,10 @@ describe('setForHearingInteractor', () => {
     });
 
     expect(
-      applicationContext.getPersistenceGateway().addHearingToCase,
+      applicationContext.getPersistenceGateway().addCaseToHearing,
     ).toHaveBeenCalled();
     expect(
-      applicationContext.getPersistenceGateway().addHearingToCase.mock
+      applicationContext.getPersistenceGateway().addCaseToHearing.mock
         .calls[0][0],
     ).toEqual(
       expect.objectContaining({
@@ -121,10 +121,10 @@ describe('setForHearingInteractor', () => {
     });
 
     expect(
-      applicationContext.getPersistenceGateway().addHearingToCase,
+      applicationContext.getPersistenceGateway().addCaseToHearing,
     ).toHaveBeenCalled();
     expect(
-      applicationContext.getPersistenceGateway().addHearingToCase.mock
+      applicationContext.getPersistenceGateway().addCaseToHearing.mock
         .calls[0][0],
     ).toEqual(
       expect.objectContaining({
